requirements.txt
affine==2.4.0
attrs==25.3.0
certifi==2025.6.15
cftime==1.6.4.post1
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
geopandas==1.1.1
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
netCDF4==1.7.2
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rasterio==1.4.3
regionmask==0.13.0
requests==2.32.4
shapely==2.1.1
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
xarray==2025.6.1
